Created additional mapping in examController to delete availabilities by studentId. Added respective service method and tests.
Created submit button which updates a students availability. The calendar stores the students availabilities as well, i.e. if they navigate away from the page and come back to it, their availabilites are still displayed in the calendar.

Modified reset button so that if a student clicks it, backend data is deleted upon confirmation by user. The calendar is also cleared. This allows students to modify their availabilites in case their plans change during busy final exam periods.

Added frontend tests as well. All are passing.