created exams page for coordinator. Coordinator can now create exam using the form provided. Coordinator can also assign the student to exam using the second form.
Updated schedule page for student to show exam assignments as well. You can test with this workflow:
add course and section
log in as coordinator
create exam for course
change studentNum for scoobert in vscode
assign scoobert to exam using the second form
log in as student
go to schedule page
verify assignments are showing up in calendar and scheduleRows.